Internal refactoring around MvcSetup and MockDispatcher
The main goal of the refactoring is to use the actual DispatcherServlet
instead of the custom MockDispatcher used till now. This was done
by introducing a DispatcherServlet sub-class, which creates
an MvcResult instance at the start of a request, stores it in a request
attribute, and populates it as the request gets executed.

The only way to initialize the DispatcherServlet is to provide it with
a WebApplicationContext containing Spring MVC infrastructure
components and controllers. As a result of this, the MvcSetup contract,
previously prepared by MockMvcBuilders and used in the MockDispatcher,
can no longer be used.

Removing the MvcSetup simplifies WebApplicationContext-based
MockMvcBuilders. However, the StandaloneMockMvcBuilder needed to be
refactored. After the refactoring it "borrows" infrastructure set up by
the MVC Java conifg WebMvcConfigurationSupport and uses it to populate
a "stub" WebApplicationContext that accepts registrations of
singletons. Hence the StandaloneMockMvcBuilder can continue to
exist and to provide an option for manual, unit test style testing
where controllers can be instantiated and configured manually with
dependencies (e.g. mock objects).

Fortunately there are no backwards compatibility issues in this change
that is unless tests did any advanced extensions of MockMvcBuilders in
which case some adjustments may be necessary.