Refine pathInfo algorithm
Before this change, the pathInfo was determined automatically. After
the change, the pathInfo can be specified, and if so the value is used
as-is, i.e. without validating that:

requestUri = contextPath + servletPath + pathInfo

The assumption is that if the test chooses to specify the pathInfo
explicitly, it is targetting a specific scenario. For example,
simulating a welcome file where for requestUri "/", the servletPath
is "/index.html"